ver Mar29th
migrated LogcatThread to the novel Event interfaces.
modified the constructor, methods to add and remove listeners, the
wait method and the response codes in the main method.

revised the initiation in TaskManager and the usage of wait in
SetupStepInterpreter accordingly.

supplemented the missed default argument for the image_format parameter
in the constructor of TaskManager.
supplemented the missed update parameter in the constructor of LogEvent.

moved the super invocation to the end of the constructor in
ScreenAnalyzerThread to correctly initiate and launch the thread.